requirements.txt
asgiref==3.4.1
Django==4.0.1
sqlparse==0.4.2
psycopg2-binary==2.9.3
flake8==4.0.1
black==21.12b0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-filter==22.1
markdown==3.3.7
click==8.0.2
redis==5.0.2
